requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
aiosqlite==0.19.0
psycopg2-binary==2.9.9

# Async HTTP - обновленная версия для совместимости
aiohttp==3.9.5

# Task Queue
celery==5.3.6
redis==5.0.1
flower==2.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.26.0

# Migrations
alembic==1.13.1

# Utilities
python-json-logger==2.0.7